Rewire enableWhen
