+++ src/components/Questionnaire/QuestionTypes/QuestionInput.tsx
@@ -9,7 +9,7 @@ import type {
   QuestionnaireResponse,
   ResponseValue,
 } from "@/types/questionnaire/form";
-import type { EnableWhen, Question } from "@/types/questionnaire/question";
+import type { Question } from "@/types/questionnaire/question";
 
 import { AllergyQuestion } from "./AllergyQuestion";
 import { BooleanQuestion } from "./BooleanQuestion";
@@ -55,54 +55,6 @@ export function QuestionInput({
     return null;
   }
 
-  const isQuestionEnabled = () => {
-    if (!question.enable_when?.length) return true;
-
-    const checkCondition = (enableWhen: EnableWhen) => {
-      const dependentValue = questionnaireResponses.find(
-        (v) => v.link_id === enableWhen.question,
-      )?.values[0];
-
-      // Early return if no dependent value exists
-      if (!dependentValue?.value) return false;
-
-      switch (enableWhen.operator) {
-        case "exists":
-          return dependentValue !== undefined && dependentValue !== null;
-        case "equals":
-          return dependentValue.value === enableWhen.answer;
-        case "not_equals":
-          return dependentValue.value !== enableWhen.answer;
-        case "greater":
-          return (
-            typeof dependentValue.value === "number" &&
-            dependentValue.value > enableWhen.answer
-          );
-        case "less":
-          return (
-            typeof dependentValue.value === "number" &&
-            dependentValue.value < enableWhen.answer
-          );
-        case "greater_or_equals":
-          return (
-            typeof dependentValue.value === "number" &&
-            dependentValue.value >= enableWhen.answer
-          );
-        case "less_or_equals":
-          return (
-            typeof dependentValue.value === "number" &&
-            dependentValue.value <= enableWhen.answer
-          );
-        default:
-          return true;
-      }
-    };
-
-    return question.enable_behavior === "any"
-      ? question.enable_when.some(checkCondition)
-      : question.enable_when.every(checkCondition);
-  };
-
   const handleAddValue = () => {
     updateQuestionnaireResponseCB({
       ...questionnaireResponse,
@@ -121,14 +73,12 @@ export function QuestionInput({
   };
 
   const renderSingleInput = (index: number = 0) => {
-    const isEnabled = isQuestionEnabled();
-
     const commonProps = {
       classes: question.styling_metadata?.classes,
       question,
       questionnaireResponse,
       updateQuestionnaireResponseCB,
-      disabled: !isEnabled || disabled,
+      disabled,
       clearError,
       index,
     };
@@ -221,7 +171,7 @@ export function QuestionInput({
             size="sm"
             onClick={handleAddValue}
             className="mt-2"
-            disabled={!isQuestionEnabled() || disabled}
+            disabled={disabled}
           >
             <CareIcon icon="l-plus" className="mr-2 h-4 w-4" />
             Add Another
@@ -230,11 +180,6 @@ export function QuestionInput({
       </div>
     );
   };
-
-  const isEnabled = isQuestionEnabled();
-  if (!isEnabled && question.disabled_display === "hidden") {
-    return null;
-  }
 
   const error = errors.find((e) => e.question_id === question.id)?.error;
 
@@ -247,7 +192,7 @@ export function QuestionInput({
         <NotesInput
           questionnaireResponse={questionnaireResponse}
           updateQuestionnaireResponseCB={updateQuestionnaireResponseCB}
-          disabled={!isEnabled}
+          disabled={disabled}
         />
       )}
     </div>
